File renaming with code refactor. Working Home page ts
