+++ src/pages/Home/Home.types.ts
@@ -1,0 +1,13 @@
+export default interface PageProps {
+  feed: IFeed;
+  getData(page: number): void;
+  resetState(): void;
+  incrementPage(): void;
+}
+
+export interface IFeed {
+  isLoading: boolean;
+  hasError: boolean;
+  data: Array<object>;
+  page: number;
+}
