+++ src/pages/Home/index.tsx
@@ -5,18 +5,26 @@ import LoadingBar from "react-top-loading-bar";
 import { ImageAndUser, LoadingCircle, ErrorPage } from "components";
 import Showcase from "components/ShowcaseContainer";
 import { DisplayArea } from "./Home.styles";
-import { getData, resetState, incrementPage } from "store/feed/feedAction";
+import {
+  getData,
+  resetState,
+  incrementPage,
+} from "store/feed/feedActionCreators";
+import PageProps from "pages/Home/Home.types";
 
-function Home(props) {
-  const ref = React.createRef();
+const Home = (props: PageProps) => {
+  const ref:any = React.createRef();
   const [showcaseLoaded, setShowcaseLoaded] = useState(false);
   const { isLoading, hasError, data } = props.feed || {};
   const hasLoadingFinish = !isLoading && showcaseLoaded;
   const hasData = !!data.length && !hasError && showcaseLoaded;
-  
+
   useEffect(() => {
     const loadingBar = ref.current;
-    isLoading ? loadingBar.continuousStart() : loadingBar.complete();
+    if (loadingBar) {
+      isLoading ? loadingBar.continuousStart() : loadingBar.complete();
+    }
+
     return function cleanup() {
       loadingBar.complete();
     };
@@ -47,9 +55,10 @@ function Home(props) {
           dataLength={data.length}
           next={props.incrementPage}
           hasMore={true}
+          loader={null}
         >
           <DisplayArea>
-            {data.map((item) => {
+            {data.map((item:any) => {
               return <ImageAndUser key={item.id} item={item} />;
             })}
           </DisplayArea>
@@ -59,9 +68,9 @@ function Home(props) {
       {hasError && <ErrorPage />}
     </>
   );
-}
+};
 
-const mapStateToProps = (state) => ({
+const mapStateToProps = (state:any) => ({
   feed: state.feed,
 });
