+++ src/store/feed/feed.enum.ts
