+++ src/store/store.type.ts
@@ -1,0 +1,6 @@
+// export default interface IStore {
+//     feed,
+// }
+
+// Finish all interfaces for each store, then add them to here
+export{};
